refactor(admin): split users list panels
Why:
- keep the Quantercise users list page focused on routing and search params
- isolate getQCUsers and search filtering from reusable list/count UI
- continue reducing mixed data and presentation code in admin routes

Verify:
- pnpm exec prettier --check apps/admin/src/app/quantercise/users/page.tsx apps/admin/src/app/quantercise/users/users-content.tsx apps/admin/src/app/quantercise/users/users-sections.tsx
- git diff --check -- apps/admin/src/app/quantercise/users/page.tsx apps/admin/src/app/quantercise/users/users-content.tsx apps/admin/src/app/quantercise/users/users-sections.tsx
- pnpm --filter @anipotts/admin typecheck
- pnpm --filter @anipotts/admin lint
- pnpm --filter @anipotts/admin build
- pnpm validate
